fix(gateway): keep instance runtime contracts acyclic
